Always add jfrog CLI to PATH
Just playing around with this action on a self hosted runner, it seems like the action succeeds the first time it is executed on a host, but fails every subsequent time. I think it is because the action only calls `core.addPath` if it actively downloaded the JFrog CLI, instead of every time the action is executed.